Update-e5d1ffb0-defe-4be6-a05e-31af180481bd

+++ src/main/java/com/bss/inc/cameraon/utils/$UI$02.java
@@ -139,7 +139,7 @@ public class $UI$02 implements StateTransitionListener, ScanningProgressCallback
     }
 
     @Override
-    public void updateProgress(InetAddress currentAddress, int runningThreads, double percentageComplete) {
+    public void updateProgress(InetAddress currentAddress, int runningThreads, int percentageComplete) {
         if (currentAddress != null) System.out.println("Trying " + ((currentAddress == null) ? "N/A" : currentAddress.getHostAddress().toString()) + " | Running Threads: " + String.valueOf(runningThreads) + " | Percent Complete: " + String.valueOf(percentageComplete) + "%");
         if (MainClass.getScanningResults().isInfoAvailable()) {
             System.out.println("Scanning took: " + String.valueOf(MainClass.getScanningResults().getScanInfo().getScanTime())+"s");
